backend/pyproject.toml
[project]
name = "eurodent2-backend"
version = "0.1.0"
description = "Sistema de gestión integral para clínica dental Eurodent 2.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "sqlalchemy>=2.0.30",
    "alembic>=1.13.1",
    "asyncpg>=0.29.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.2.1",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.0.0",
    "psycopg2-binary>=2.9.9",
    "python-multipart>=0.0.9",
    "httpx>=0.27.0",
    "reportlab>=4.2.0",
    "qrcode[pil]>=7.4.2",
    "Pillow>=10.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.6",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["app"]
